requirements.txt
cyclone
pygithub
pyyaml
requests
twisted
